fastapi_app/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.2.0
faiss-cpu==1.7.4
torch==2.1.2
clip @ git+https://github.com/openai/CLIP.git
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-dotenv==1.0.0 